refactoring movie bloc
