+++ lib/domain/blocs/movie/movie_details_new_event.dart
